feat: create data mock folder
